I fixed or silenced all warnings. The code is now compiling with `-Werror`
I also changed warning flags to remove broken warnings from gcc

Even bigger changes:
I deleted the following folders and subystems:
-elf_loader
-storage (including all of the ATA driver)
-string_struct (which was essentially empty)
-float (which was mostly broken and incomplete with most code being commented out already)

+++ src/drivers/vga_driver/vga_driver.h
@@ -49,11 +49,11 @@ struct vga_driver_context {
 
 //these two functions are correct by inspection:
 inline uint8_t vga_entry_color(const enum vga_color fg, const enum vga_color bg) {
-    return fg | bg << 4u;
+    return (uint8_t) (fg | bg << 4u);
 }
 
 inline uint16_t vga_entry(const char uc, const enum vga_color color) {
-    return (uint16_t) uc | (uint16_t) color << 8u;
+    return (uint16_t) ((uint16_t) uc | (uint16_t) color << 8u);
 }
